Create 'pesajes' table and related files
A new migration file has been added to create a 'pesajes' table in the database. The corresponding model, factory, and seeder files for the 'pesajes' table have also been created. The model file 'Pesaje.php' includes the necessary properties and the factory and seeder files are currently placeholders ready for future data generation and seeding needs.